+++ zhmcclient/_certificates.py
@@ -34,7 +34,7 @@ import copy
 from ._logging import logged_api_call
 from ._manager import BaseManager
 from ._resource import BaseResource
-from ._utils import matches_filters, divide_filter_args, RC_CERTIFICATE
+from ._utils import RC_CERTIFICATE
 
 __all__ = ['CertificateManager', 'Certificate']
 
@@ -87,7 +87,9 @@ class CertificateManager(BaseManager):
         return self._parent
 
     @logged_api_call
-    def list(self, full_properties=False, filter_args=None):
+    # pylint: disable=arguments-differ
+    def list(self, full_properties=False, filter_args=None,
+             additional_properties=None):
         """
         List the certificates defined in the HMC.
 
@@ -111,6 +113,10 @@ class CertificateManager(BaseManager):
             `None` causes no filtering to happen, i.e. all resources are
             returned.
 
+          additional_properties (list of string):
+            List of property names that are to be returned in addition to the
+            default properties.
+
         Returns:
 
           : A list of :class:`~zhmcclient.Certificate` objects.
@@ -122,38 +128,11 @@ class CertificateManager(BaseManager):
           :exc:`~zhmcclient.AuthError`
           :exc:`~zhmcclient.ConnectionError`
         """
-
-        resource_obj_list = []
-
-        resource_obj = self._try_optimized_lookup(filter_args)
-        if resource_obj:
-            resource_obj_list.append(resource_obj)
-            # It already has full properties
-        else:
-            query_parms, client_filters = divide_filter_args(
-                self._query_props, filter_args)
-
-            resources_name = 'certificates'
-            uri = '/api/{}{}'.format(resources_name, query_parms)
-
-            result = self.session.get(uri)
-            if result:
-                props_list = result[resources_name]
-                for props in props_list:
-
-                    resource_obj = self.resource_class(
-                        manager=self,
-                        uri=props[self._uri_prop],
-                        name=props.get(self._name_prop, None),
-                        properties=props)
-
-                    if matches_filters(resource_obj, client_filters):
-                        resource_obj_list.append(resource_obj)
-                        if full_properties:
-                            resource_obj.pull_full_properties()
-
-        self._name_uri_cache.update_from(resource_obj_list)
-        return resource_obj_list
+        result_prop = 'certificates'
+        list_uri = '/api/certificates'
+        return self._list_with_operation(
+            list_uri, result_prop, full_properties, filter_args,
+            additional_properties)
 
     @logged_api_call
     def import_certificate(self, cpc, properties):