+++ zhmcclient/_cpc.py
@@ -66,7 +66,7 @@ from ._virtual_switch import VirtualSwitchManager
 from ._capacity_group import CapacityGroupManager
 from ._logging import logged_api_call
 from ._exceptions import ParseError, ConsistencyError
-from ._utils import get_features, matches_filters, divide_filter_args, \
+from ._utils import get_features, \
     RC_CPC, RC_ADAPTER, RC_HBA, RC_NIC, RC_PARTITION, \
     RC_NETWORK_PORT, RC_STORAGE_PORT, RC_STORAGE_TEMPLATE, RC_STORAGE_GROUP, \
     RC_STORAGE_TEMPLATE_VOLUME, RC_STORAGE_VOLUME, RC_VIRTUAL_FUNCTION, \
@@ -257,43 +257,10 @@ class CpcManager(BaseManager):
           :exc:`~zhmcclient.AuthError`
           :exc:`~zhmcclient.ConnectionError`
         """
-        resource_obj_list = []
-        if self.auto_update_enabled() and not self.auto_update_needs_pull():
-            for resource_obj in self.list_resources_local():
-                if matches_filters(resource_obj, filter_args):
-                    resource_obj_list.append(resource_obj)
-        else:
-            resource_obj = self._try_optimized_lookup(filter_args)
-            if resource_obj:
-                resource_obj_list.append(resource_obj)
-                # It already has full properties
-            else:
-                query_parms, client_filters = divide_filter_args(
-                    self._query_props, filter_args)
-
-                resources_name = 'cpcs'
-                uri = '/api/{}{}'.format(resources_name, query_parms)
-
-                result = self.session.get(uri)
-                if result:
-                    props_list = result[resources_name]
-                    for props in props_list:
-
-                        resource_obj = self.resource_class(
-                            manager=self,
-                            uri=props[self._uri_prop],
-                            name=props.get(self._name_prop, None),
-                            properties=props)
-
-                        if matches_filters(resource_obj, client_filters):
-                            resource_obj_list.append(resource_obj)
-                            if full_properties:
-                                resource_obj.pull_full_properties()
-
-            self.add_resources_local(resource_obj_list)
-
-        self._name_uri_cache.update_from(resource_obj_list)
-        return resource_obj_list
+        result_prop = 'cpcs'
+        list_uri = '/api/cpcs'
+        return self._list_with_operation(
+            list_uri, result_prop, full_properties, filter_args, None)
 
 
 class Cpc(BaseResource):