+++ zhmcclient/_lpar.py
@@ -32,9 +32,9 @@ import copy
 
 from ._manager import BaseManager
 from ._resource import BaseResource
-from ._exceptions import StatusTimeout, HTTPError
+from ._exceptions import StatusTimeout
 from ._logging import logged_api_call
-from ._utils import matches_filters, divide_filter_args, RC_LOGICAL_PARTITION
+from ._utils import RC_LOGICAL_PARTITION
 
 __all__ = ['LparManager', 'Lpar']
 
@@ -144,51 +144,10 @@ class LparManager(BaseManager):
           :exc:`~zhmcclient.AuthError`
           :exc:`~zhmcclient.ConnectionError`
         """
-        resource_obj_list = []
-        if self.auto_update_enabled() and not self.auto_update_needs_pull():
-            for resource_obj in self.list_resources_local():
-                if matches_filters(resource_obj, filter_args):
-                    resource_obj_list.append(resource_obj)
-        else:
-            resource_obj = self._try_optimized_lookup(filter_args)
-            if resource_obj:
-                resource_obj_list.append(resource_obj)
-                # It already has full properties
-            else:
-                query_parms, client_filters = divide_filter_args(
-                    self._query_props, filter_args)
-
-                resources_name = 'logical-partitions'
-                uri = '{}/{}{}'.format(self.cpc.uri, resources_name,
-                                       query_parms)
-
-                try:
-                    result = self.session.get(uri)
-                except HTTPError as exc:
-                    if exc.http_status == 404 and exc.reason == 1:
-                        # Unlike other list operations, "List Logical Partitions
-                        # of CPC" fails with 404.1 "ERROR: found no Images" if
-                        # no LPAR matches the filters in the query parms.
-                        result = []
-                if result:
-                    props_list = result[resources_name]
-                    for props in props_list:
-
-                        resource_obj = self.resource_class(
-                            manager=self,
-                            uri=props[self._uri_prop],
-                            name=props.get(self._name_prop, None),
-                            properties=props)
-
-                        if matches_filters(resource_obj, client_filters):
-                            resource_obj_list.append(resource_obj)
-                            if full_properties:
-                                resource_obj.pull_full_properties()
-
-            self.add_resources_local(resource_obj_list)
-
-        self._name_uri_cache.update_from(resource_obj_list)
-        return resource_obj_list
+        result_prop = 'logical-partitions'
+        list_uri = '{}/logical-partitions'.format(self.cpc.uri)
+        return self._list_with_operation(
+            list_uri, result_prop, full_properties, filter_args, None)
 
 
 class Lpar(BaseResource):