+++ zhmcclient/_partition.py
@@ -42,7 +42,7 @@ from ._nic import NicManager
 from ._hba import HbaManager
 from ._virtual_function import VirtualFunctionManager
 from ._logging import logged_api_call
-from ._utils import matches_filters, divide_filter_args, RC_PARTITION
+from ._utils import RC_PARTITION
 
 __all__ = ['PartitionManager', 'Partition']
 
@@ -98,7 +98,9 @@ class PartitionManager(BaseManager):
         return self._parent
 
     @logged_api_call
-    def list(self, full_properties=False, filter_args=None):
+    # pylint: disable=arguments-differ
+    def list(self, full_properties=False, filter_args=None,
+             additional_properties=None):
         """
         List the Partitions in this CPC.
 
@@ -143,6 +145,12 @@ class PartitionManager(BaseManager):
             `None` causes no filtering to happen, i.e. all resources are
             returned.
 
+          additional_properties (list of string):
+            List of property names that are to be returned in addition to the
+            default properties.
+
+            This parameter requires HMC 2.16.0 or higher.
+
         Returns:
 
           : A list of :class:`~zhmcclient.Partition` objects.
@@ -154,44 +162,11 @@ class PartitionManager(BaseManager):
           :exc:`~zhmcclient.AuthError`
           :exc:`~zhmcclient.ConnectionError`
         """
-        resource_obj_list = []
-        if self.auto_update_enabled() and not self.auto_update_needs_pull():
-            for resource_obj in self.list_resources_local():
-                if matches_filters(resource_obj, filter_args):
-                    resource_obj_list.append(resource_obj)
-        else:
-            resource_obj = self._try_optimized_lookup(filter_args)
-            if resource_obj:
-                resource_obj_list.append(resource_obj)
-                # It already has full properties
-            else:
-                query_parms, client_filters = divide_filter_args(
-                    self._query_props, filter_args)
-
-                resources_name = 'partitions'
-                uri = '{}/{}{}'.format(self.cpc.uri, resources_name,
-                                       query_parms)
-
-                result = self.session.get(uri)
-                if result:
-                    props_list = result[resources_name]
-                    for props in props_list:
-
-                        resource_obj = self.resource_class(
-                            manager=self,
-                            uri=props[self._uri_prop],
-                            name=props.get(self._name_prop, None),
-                            properties=props)
-
-                        if matches_filters(resource_obj, client_filters):
-                            resource_obj_list.append(resource_obj)
-                            if full_properties:
-                                resource_obj.pull_full_properties()
-
-            self.add_resources_local(resource_obj_list)
-
-        self._name_uri_cache.update_from(resource_obj_list)
-        return resource_obj_list
+        result_prop = 'partitions'
+        list_uri = '{}/partitions'.format(self.cpc.uri)
+        return self._list_with_operation(
+            list_uri, result_prop, full_properties, filter_args,
+            additional_properties)
 
     @logged_api_call
     def create(self, properties):